Fix styling of navigation drawer
