+++ asreview/webapp/src/App.css
@@ -31,16 +31,6 @@
   font-size: 140% !important;
 }
 
-.navDrawerSelected::after {
-  top: 0;
-  left: 0;
-  width: 4px;
-  height: 100%;
-  content: " ";
-  position: absolute;
-  background-color: #91620b;
-}
-
 .paperProjectsInReview {
   background-color: #fff7cd !important;
 }
