+++ asreview/webapp/src/Components/index.js
@@ -1,6 +1,8 @@
 export { default as AppBarWithinDialog } from "./AppBarWithinDialog";
 export { default as DashboardStatsPaper } from "./DashboardStatsPaper";
 export { default as DialogTitleWithClose } from "./DialogTitleWithClose";
+export { default as DrawerItem } from "./DrawerItem";
+export { default as DrawerItemContainer } from "./DrawerItemContainer";
 export { default as ExportDialog } from "./ExportDialog";
 export { default as Header } from "./Header";
 export { default as NavigationDrawer } from "./NavigationDrawer";
@@ -8,4 +10,3 @@ export { default as OpenInNewIconStyled } from "./OpenInNewIconStyled";
 export { default as ProjectImportDialog } from "./ProjectImportDialog";
 export { default as ProjectTable } from "./ProjectTable";
 export { default as QuickTourDialog } from "./QuickTourDialog";
-export { default as StyledDrawerItem } from "./StyledDrawerItem";
